requirements.txt
pillow
glob2